+++ src/check.ts
@@ -1,16 +1,30 @@
 // generated by src/check.schema.ts
 
 export interface CheckTaskArgs {
-	_: string[];
 	typecheck?: boolean;
+	/**
+	 * opt out of typechecking
+	 */
 	'no-typecheck'?: boolean;
 	test?: boolean;
+	/**
+	 * opt out of running tests
+	 */
 	'no-test'?: boolean;
 	gen?: boolean;
+	/**
+	 * opt out of gen check
+	 */
 	'no-gen'?: boolean;
 	format?: boolean;
+	/**
+	 * opt out of format check
+	 */
 	'no-format'?: boolean;
 	lint?: boolean;
+	/**
+	 * opt out of linting
+	 */
 	'no-lint'?: boolean;
 }
